requirements.txt
colpali-engine>=0.3.15
qdrant-client>=1.10.0
pdf2image>=1.17.0
pillow>=10.0.0
torch>=2.0.0
transformers>=5.3.0,<5.4.0
streamlit>=1.35.0
python-dotenv>=1.0.0
datasets>=2.20.0
huggingface_hub>=0.23.0
tqdm>=4.66.0
requests>=2.31.0
